+++ java-client/src/main/java/io/sphere/client/model/EmptyReference.java
@@ -50,9 +50,31 @@ public class EmptyReference<T> extends Reference<T> {
     @Override public String getTypeId() throws ReferenceException { throw emptyReferenceException(); }
 
     @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof EmptyReference)) return false;
+
+        EmptyReference that = (EmptyReference) o;
+
+        if (!fieldName.equals(that.fieldName)) return false;
+
+        return true;
+    }
+
+    @Override
+    public int hashCode() {
+        return fieldName.hashCode();
+    }
+
+    @Override
     public String toString() {
         return "EmptyReference{" +
                 "fieldName='" + fieldName + '\'' +
                 '}';
+    }
+
+    @Override
+    public ReferenceId<T> toReferenceIdOrNull() {
+        return null;
     }
 }