+++ java-client/src/main/java/io/sphere/client/shop/InventoryService.java
@@ -12,12 +12,13 @@ import org.joda.time.DateTime;
  * Sphere HTTP API for retrieving product inventory information.
  */
 public interface InventoryService {
-    /** Creates a new InventoryEntry that does not belong to a SupplyChannel.
+    /** Creates a new InventoryEntry that does not belong to a Channel.
      *
      * Throws a io.sphere.client.exceptions.DuplicateSkuException if the sku is already present.
      **/
     CommandRequest<InventoryEntry> createInventoryEntry(String sku, long quantityOnStock);
-    /** Creates a new InventoryEntry that does not belong to a SupplyChannel.
+
+    /** Creates a new InventoryEntry that does not belong to a Channel.
      *
      * Throws a io.sphere.client.exceptions.DuplicateSkuException if the sku is already present.
      *
@@ -25,9 +26,28 @@ public interface InventoryService {
      * @param restockableInDays optional the time difference in days to get the item in stock again.
      **/
     CommandRequest<InventoryEntry> createInventoryEntry(String sku, long quantityOnStock, Long restockableInDays, DateTime expectedDelivery);
-    /** Fetches the InventoryEntry by sku which does not have any supply channel */
+
+    /** Creates a new InventoryEntry that does belong to a Channel.
+     *
+     * Throws a io.sphere.client.exceptions.DuplicateSkuException if the sku is already present.
+     * Throws a io.sphere.client.SphereError.ResourceNotFound if the Channel does not exist.
+     *
+     * @param expectedDelivery optional expected date for the delivery of the item.
+     * @param restockableInDays optional the time difference in days to get the item in stock again.
+     **/
+    CommandRequest<InventoryEntry> createInventoryEntry(String sku, long quantityOnStock, Long restockableInDays, DateTime expectedDelivery, String channelId);
+
+    /** Fetches the InventoryEntry by sku which does not have any channel. */
     FetchRequest<InventoryEntry> bySku(String sku);
+
+    /** Fetches the InventoryEntry by sku which does have a channel. */
+    FetchRequest<InventoryEntry> bySku(String sku, String channelId);
+
+    /** Searches all InventoryEntry objects that have the given sku. Finds the objects with and without Channel. */
+    QueryRequest<InventoryEntry> queryBySku(String sku);
+
     /** Queries inventory entries. */
     QueryRequest<InventoryEntry> query();
+
     CommandRequest<InventoryEntry> updateInventoryEntry(VersionedId id, InventoryEntryUpdate update);
 }