+++ java-client/src/main/java/io/sphere/client/shop/SphereClient.java
@@ -32,6 +32,7 @@ final public class SphereClient {
     private final ShippingMethodService shippingMethodService;
     private final TaxCategoryService    taxCategoryService;
     private final CustomObjectService   customObjectService;
+    private final ChannelService channelService;
 
     /** Creates an instance of SphereClient.
      *
@@ -50,7 +51,8 @@ final public class SphereClient {
                         InventoryService inventoryService,
                         ShippingMethodService shippingMethodService,
                         TaxCategoryService taxCategoryService,
-                        CustomObjectService customObjectService) {
+                        CustomObjectService customObjectService,
+                        ChannelService channelService) {
         this.config            = config;
         this.httpClient        = httpClient;
         this.clientCredentials = clientCredentials;
@@ -65,6 +67,7 @@ final public class SphereClient {
         this.shippingMethodService = shippingMethodService;
         this.taxCategoryService    = taxCategoryService;
         this.customObjectService = customObjectService;
+        this.channelService = channelService;
     }
 
     /** Creates an instance of SphereClient. */
@@ -97,7 +100,8 @@ final public class SphereClient {
             new InventoryServiceImpl(requestFactory, projectEndpoints),
             new ShippingMethodServiceImpl(requestFactory, projectEndpoints),
             new TaxCategoryServiceImpl(requestFactory, projectEndpoints),
-            new CustomObjectServiceImpl(requestFactory, projectEndpoints)
+            new CustomObjectServiceImpl(requestFactory, projectEndpoints),
+            new ChannelServiceImpl(requestFactory, projectEndpoints)
         );
     }
 
@@ -147,4 +151,7 @@ final public class SphereClient {
 
     /** Provides access to custom objects. */
     public CustomObjectService customObjects() { return customObjectService; }
+
+    /** Provides access to supply channels. */
+    public ChannelService channels() { return channelService; }
 }