+++ java-client/src/main/java/io/sphere/client/shop/model/ChannelRoles.java
@@ -1,0 +1,5 @@
+package io.sphere.client.shop.model;
+
+public enum ChannelRoles {
+    InventorySupply, Primary
+}