+++ java-client/src/main/java/io/sphere/client/shop/model/InventoryEntry.java
@@ -1,6 +1,7 @@
 package io.sphere.client.shop.model;
 
 import com.google.common.base.Optional;
+import io.sphere.client.model.Reference;
 import io.sphere.client.model.VersionedId;
 import io.sphere.internal.util.Iso8601JsonSerializer;
 import net.jcip.annotations.Immutable;
@@ -22,6 +23,7 @@ public class InventoryEntry {
     private Integer restockableInDays;
     @JsonSerialize(using = Iso8601JsonSerializer.class)
     private DateTime expectedDelivery;
+    Reference<Channel> channel = Channel.emptyReference();
 
     // for JSON deserializer
     protected InventoryEntry() {}
@@ -49,6 +51,10 @@ public class InventoryEntry {
         return expectedDelivery;
     }
 
+    public Reference<Channel> getChannel() {
+        return channel;
+    }
+
     @Override
     public String toString() {
         return "InventoryEntry{" +
@@ -59,6 +65,7 @@ public class InventoryEntry {
                 ", availableQuantity=" + availableQuantity +
                 ", restockableInDays=" + restockableInDays +
                 ", expectedDelivery=" + expectedDelivery +
+                ", channel=" + channel +
                 '}';
     }
 }