+++ java-client/src/test/scala/MockSphereClient.scala
@@ -31,7 +31,8 @@ object MockSphereClient {
     inventoryResponse:  FakeResponse = nullResponse("Inventory"),
     shippingMethodsResponse: FakeResponse = nullResponse("ShippingMethods"),
     taxCategoriesResponse: FakeResponse = nullResponse("TaxCategories"),
-    customObjectResponse: FakeResponse = nullResponse("CustomObjects")): SphereClient =
+    customObjectResponse: FakeResponse = nullResponse("CustomObjects"),
+    supplyChannelResponse: FakeResponse = nullResponse("Channel")): SphereClient =
   {
     val categoryTree = CategoryTreeImpl.createAndBeginBuildInBackground(new CategoriesImpl(reqFactory(categoriesResponse), endpoints), Locale.ENGLISH)
     new SphereClient(
@@ -47,7 +48,8 @@ object MockSphereClient {
       new InventoryServiceImpl(reqFactory(inventoryResponse), endpoints),
       new ShippingMethodServiceImpl(reqFactory(shippingMethodsResponse), endpoints),
       new TaxCategoryServiceImpl(reqFactory(taxCategoriesResponse), endpoints),
-      new CustomObjectServiceImpl(reqFactory(customObjectResponse), endpoints)
+      new CustomObjectServiceImpl(reqFactory(customObjectResponse), endpoints),
+      new ChannelServiceImpl(reqFactory(customObjectResponse), endpoints)
    )
   }
 }