+++ play-sdk/app/sphere/internal/InventoryServiceAdapter.java
@@ -2,11 +2,7 @@ package sphere.internal;
 
 import net.jcip.annotations.Immutable;
 import sphere.InventoryService;
-import io.sphere.client.SphereResult;
-import io.sphere.client.model.VersionedId;
 import io.sphere.client.shop.model.InventoryEntry;
-import io.sphere.client.shop.model.InventoryEntryUpdate;
-import play.libs.F;
 import sphere.FetchRequest;
 import sphere.QueryRequest;
 import sphere.util.Async;
@@ -31,17 +27,17 @@ public class InventoryServiceAdapter implements InventoryService {
     }
 
     @Override
+    public FetchRequest<InventoryEntry> bySku(String sku, String channelId) {
+        return Async.adapt(service.bySku(sku, channelId));
+    }
+
+    @Override
+    public QueryRequest<InventoryEntry> queryBySku(String sku) {
+        return Async.adapt(service.queryBySku(sku));
+    }
+
+    @Override
     public QueryRequest<InventoryEntry> query() {
         return Async.adapt(service.query());
-    }
-
-    @Override
-    public InventoryEntry updateInventoryEntry(VersionedId id, InventoryEntryUpdate update) {
-        return service.updateInventoryEntry(id, update).execute();
-    }
-
-    @Override
-    public F.Promise<SphereResult<InventoryEntry>> updateInventoryEntryAsync(VersionedId id, InventoryEntryUpdate update) {
-        return Async.execute(service.updateInventoryEntry(id, update));
     }
 }